feat: update tests to use the new fixture system
