+++ core/integration/testdata/modules/typescript/entrypoint-class-name-collision/dagger.json
@@ -1,0 +1,7 @@
+{
+  "name": "test",
+  "engineVersion": "latest",
+  "sdk": {
+    "source": "typescript"
+  }
+}
