+++ core/integration/testdata/modules/typescript/entrypoint-class-name-collision/package.json
@@ -1,0 +1,6 @@
+{
+  "type": "module",
+  "dependencies": {
+    "typescript": "^5.5.4"
+  }
+}
