+++ core/integration/testdata/modules/typescript/entrypoint-class-name-collision/src/index.ts
@@ -1,0 +1,17 @@
+import { func, object } from "@dagger.io/dagger"
+
+@object()
+export class Test {
+  @func()
+  makeContext(): Context {
+    return new Context()
+  }
+}
+
+@object()
+export class Context {
+  @func()
+  value(): string {
+    return "ok"
+  }
+}
