+++ core/integration/testdata/modules/typescript/entrypoint-default-export/dagger.json
@@ -1,0 +1,7 @@
+{
+  "name": "test",
+  "engineVersion": "latest",
+  "sdk": {
+    "source": "typescript"
+  }
+}
