+++ core/integration/testdata/modules/typescript/entrypoint-default-export/src/index.ts
@@ -1,0 +1,9 @@
+import { func, object } from "@dagger.io/dagger"
+
+@object()
+export default class Test {
+  @func()
+  hello(): string {
+    return "hello"
+  }
+}
