+++ core/integration/testdata/modules/typescript/entrypoint-variadic-args/dagger.json
@@ -1,0 +1,7 @@
+{
+  "name": "test",
+  "engineVersion": "latest",
+  "sdk": {
+    "source": "typescript"
+  }
+}
