+++ core/integration/testdata/modules/typescript/entrypoint-variadic-args/src/index.ts
@@ -1,0 +1,9 @@
+import { func, object } from "@dagger.io/dagger"
+
+@object()
+export class Test {
+  @func()
+  words(...words: string[]): string {
+    return words.length.toString()
+  }
+}
